Move APY and authz checks to cosmos.directory
