+++ package.json
@@ -37,8 +37,7 @@
     "start": "npx parcel src/index.html",
     "prebuild": "rm -rf dist/",
     "build": "npx parcel build src/index.html",
-    "autostake": "node scripts/autostake.mjs",
-    "checkAuthz": "node scripts/checkAuthz.mjs"
+    "autostake": "node scripts/autostake.mjs"
   },
   "eslintConfig": {
     "extends": [
