+++ src/utils/Network.mjs
@@ -2,7 +2,6 @@ import _ from 'lodash'
 import { multiply, pow, format, bignumber } from 'mathjs'
 import QueryClient from './QueryClient.mjs'
 import SigningClient from './SigningClient.mjs'
-import ApyClient from '../ApyClient.mjs'
 import Operator from './Operator.mjs'
 import Chain from './Chain.mjs'
 import CosmosDirectory from './CosmosDirectory.mjs'
@@ -48,6 +47,7 @@ class Network {
     this.decimals = this.chain.decimals
     this.image = this.chain.image
     this.coinGeckoId = this.chain.coinGeckoId
+    this.estimatedApr = this.chain.estimatedApr
     this.authzSupport = this.chain.authzSupport
     const defaultGasPrice = format(bignumber(multiply(0.000000025, pow(10, this.decimals))), { notation: 'fixed' }) + this.denom
     this.gasPrice = this.data.gasPrice || defaultGasPrice
@@ -58,15 +58,31 @@ class Network {
   async connect() {
     try {
       this.queryClient = await QueryClient(this.chain.chainId, this.rpcUrl, this.restUrl)
-      this.apyClient = ApyClient(this.chain, this.queryClient.rpcUrl, this.queryClient.restUrl)
       this.restUrl = this.queryClient.restUrl
       this.rpcUrl = this.queryClient.rpcUrl
-      this.getApy = this.apyClient.getApy
       this.connected = this.queryClient.connected
     } catch (error) {
       console.log(error)
       this.connected = false
     }
+  }
+
+  async getApy(validators, operators){
+    const chainApr = this.chain.estimatedApr
+    let validatorApy = {};
+    for (const [address, validator] of Object.entries(validators)) {
+      if(validator.jailed || validator.status !== 'BOND_STATUS_BONDED'){
+        validatorApy[address] = 0
+      }else{
+        const commission = validator.commission.commission_rates.rate
+        const operator = operators.find((el) => el.address === address)
+        const periodPerYear = operator && this.chain.authzSupport ? operator.runsPerDay() * 365 : 1;
+        const realApr = chainApr * (1 - commission);
+        const apy = (1 + realApr / periodPerYear) ** periodPerYear - 1;
+        validatorApy[address] = apy;
+      }
+    }
+    return validatorApy;
   }
 
   signingClient(wallet, key, gasPrice) {
